Fix：ProcedureEditor

+++ SKFramework/Runtime/Resources/AssetBundleUtility.cs
@@ -1,11 +1,12 @@
 /*============================================================
  * SKFramework
- * Copyright © 2019-2025 Zhang Shoukun. All rights reserved.
+ * Copyright © 2019-2026 Zhang Shoukun. All rights reserved.
  * Feedback: mailto:136512892@qq.com
  *============================================================*/
 
 #if UNITY_EDITOR
 using System;
+using System.Linq;
 using System.Text.RegularExpressions;
 
 using UnityEditor;
@@ -79,7 +80,8 @@ namespace SK.Framework.Resource
                 }
                 return false;
             }
-            AssetDatabase.RemoveAssetBundleName(assetBundleName, true);
+            if (AssetDatabase.GetAllAssetBundleNames().Contains(assetBundleName))
+                AssetDatabase.RemoveAssetBundleName(assetBundleName, true);
             return true;
         }
     }
